Add docs and rename
